055: Extracted mouse/keyboard callbacks to mixins, fixed frame rate control, extracted Processing widget, extracted Sketch class.
